Extract helper functions and constants for easier mocking
